% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{.COMDate_to_POSIX}
\alias{.COMDate_to_POSIX}
\title{COMDate to POSIX
from https://github.com/aecoleman/extrospectr/}
\usage{
.COMDate_to_POSIX(x)
}
\arguments{
\item{x}{COMDate object}
}
\value{
POSIXct
}
\description{
COMDate to POSIX
from https://github.com/aecoleman/extrospectr/
}
